% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ADtest}
\alias{AD_pval}
\alias{ADtest}
\title{Anderson-Darling test of uniformity}
\usage{
ADtest(x)

AD_pval(n, z)
}
\arguments{
\item{x}{a numeric vector (preferably longer than 7 elements)}

\item{n}{length of original sequence}

\item{z}{Anderson-Darling test statistic}
}
\value{
an object of class \code{htest} containing the test statistic and p-value for a test against a null hypothesis of uniformity

Anderson-Darling p-value
}
\description{
Anderson-Darling test of uniformity

Anderson-Darling p-value
}
\examples{
u <- c(.0392,.0884,.260,.310,.454,.644,.797,.813,.921,.960)
(t1 <- ADtest(u))
AD_pval(length(u),t1$statistic)
}
\references{
George Marsaglia and John Marsaglia (2004) "Evaluating the Anderson-Darling distribution", \emph{J. Statistical Software} vol. 9, issue i02: \url{http://www.jstatsoft.org/v09/i02/paper}
}

